fixed problem to Frechet derivative lower bound.
Frechet derivatives of travel time with respect to grid velocities had a fixed floor value of 1.0E-6. For certain problems, e.g. in Cartesian co-ordinates where velocities are in m/s and have values of 10^3, this can cause derivatives of travel time with respect to slowness to have a floor of ~1.0, since dt/ds = -(v**2)*dt/dv. Hence slowness derivative matrix can be set to zero which causes flow on errors in tomography examples. Have replaced the floor with 1.0E-10 for now. This may require further work.